+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Fix bolt issue when the catalog is referenced
 - Fix bolt issue when scriptcompiler scope is used
   and the debugger tries to expand the resources
   in the catalog.